Add reproducer for InternalExecRuntime connection leak
It is possible for another thread to cancel the HTTP request while the
calling thread is leasing a connection from the connection pool. When
this happens, it is vital that any connection pool entries that were
allocated be released, in order to avoid a connection leak. It turns out
that the call to `connRequest#cancel`, in the cancellation check
immediately after the call to `manager#lease`, does not do this: if the
`LeaseRequest` (which is essentially a type of future) is already
complete, the call to `connRequest#cancel` will not release the
connection pool entry.